Add InputProcessorLogic

+++ src/main/java/IOProcessor/IOutputProcessor.java
@@ -2,6 +2,6 @@ package IOProcessor;
 
 public interface IOutputProcessor {
     void showIntergalacticToArabicConversion(String intergalactic, int arabic);
-    void showCreditValueOfMetal(String intergalactic, String metal, int Credits);
+    void showCreditValueOfMetal(String intergalactic, String metal, double Credits);
     void showErrorMessage();
 }
